updates cli
