+++ .github/workflows/docs.yml
@@ -4,7 +4,21 @@ on:
   push:
     branches:
       - develop
-      - master
+    paths:
+      - 'docs/**'
+      - 'mkdocs.yaml'
+      - './github/workflows/docs.yml'
+  pull_request:
+    branches:
+      - develop
+    types:
+      - opened
+      - synchronize
+    paths:
+      - 'docs/**'
+      - 'mkdocs.yaml'
+      - './github/workflows/docs.yml'
+
   schedule:
     - cron: '37 23 * * 2'
 
@@ -18,59 +32,40 @@ concurrency:
   cancel-in-progress: false
 
 jobs:
-#  changes:
-#    runs-on: ubuntu-latest
-#    timeout-minutes: 1
-#    defaults:
-#      run:
-#        shell: bash
-#    outputs:
-#      docs: ${{ steps.changed_files.outputs.docs }}
-#    steps:
-#      - name: Checkout code
-#        uses: actions/checkout@v4.1.7
-#      - id: changed_files
-#        name: Check for file changes
-#        uses: dorny/paths-filter@0bc4621a3135347011ad047f9ecf449bf72ce2bd # v3.0.0
-#        with:
-#          base: ${{ github.ref }}
-#          token: ${{ github.token }}
-#          filters: .github/file-filters.yml
+  changes:
+    runs-on: ubuntu-latest
+    timeout-minutes: 1
+    defaults:
+      run:
+        shell: bash
+    outputs:
+      docs: ${{ steps.changed_files.outputs.docs }}
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+      - id: changed_files
+        name: Check for file changes
+        uses: dorny/paths-filter@v3
+        with:
+          base: ${{ github.ref }}
+          token: ${{ github.token }}
+          filters: .github/file-filters.yml
   generate:
     name: Generate
-#    if: needs.changes.outputs.docs == 'true'
-#    needs: changes
+    if: needs.changes.outputs.docs == 'true'
+    needs: changes
     runs-on: ubuntu-latest
     env:
       PYTHONPATH: src/
     steps:
       - uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: '3.12'
+      - run: pip install tox
+      - name: Check Python Package structure
+        run: tox -e docs -- -d docs-output
 
-      - uses: yezz123/setup-uv@v4
-
-      - uses: actions/cache/restore@v4
-        id: restore-cache
-        with:
-          path: .venv
-          key: ${{ runner.os }}-${{ hashFiles('**/uv.lock') }}
-
-      - name: Install dependencies
-        run: uv sync --only-group docs
-
-      - name: Build Doc
-        run: .venv/bin/mkdocs build -d ./docs-output
-
-      - uses: actions/cache/save@v4
-        id: cache
-        if: always() && steps.restore-cache.outputs.cache-hit != 'true'
-        with:
-          path: .venv
-          key: ${{ runner.os }}-${{ hashFiles('**/uv.lock') }}
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
@@ -79,6 +74,7 @@ jobs:
   # Deployment job
   deploy:
     needs: generate
+    if: github.event_name == 'push' && github.ref == 'refs/heads/develop'
     environment:
       name: github-pages
       url: ${{ steps.deployment.outputs.page_url }}
