+++ .github/workflows/lint.yml
@@ -1,17 +1,16 @@
 name: Lint
 on:
+  workflow_dispatch:
   push:
     branches:
       - develop
-      - master
-      - staging
-      - release/*
-      - feature/*
-      - bugfix/*
-      - hotfix/*
-#  pull_request:
-#    branches: [develop, master]
-#    types: [synchronize, opened, reopened, ready_for_review]
+  pull_request:
+    types: [ synchronize, opened, reopened, ready_for_review ]
+    paths:
+      - 'src/**'
+      - 'tests/**'
+      - '*.ini'
+      - '*.toml'
 
 defaults:
   run:
@@ -34,18 +33,20 @@ jobs:
     timeout-minutes: 3
     outputs:
       lint: ${{ steps.changes.outputs.lint }}
-      docker: ${{ steps.changes.outputs.docker_base }}
+      mypy: ${{ steps.changes.outputs.mypy }}
+      package: ${{ steps.changes.outputs.package }}
     steps:
       - run: git config --global --add safe.directory $(realpath .)
-      - uses: actions/checkout@b4ffde65f46336ab88eb53be808477a3936bae11 # v4.1.1
+      - uses: actions/checkout@v4
       - id: changes
         name: Check for backend file changes
-        uses: dorny/paths-filter@0bc4621a3135347011ad047f9ecf449bf72ce2bd # v3.0.0
+        uses: dorny/paths-filter@v3
         with:
           base: ${{ github.ref }}
           token: ${{ github.token }}
           filters: .github/file-filters.yml
-  ruff:
+
+  lint:
     needs: changes
     runs-on: ubuntu-latest
     if: github.event.pull_request.draft == false && needs.changes.outputs.lint
@@ -53,11 +54,34 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: '3.12'
-      - name: Install requirements
-        run: pip install ruff
-      - name: Check syntax
-        run: ruff check --output-format concise src/
+          python-version: '3.13.1'
+      - run: pip install tox
+      - name: Static Code checks
+        run: tox -e lint
 
-      - name: Warnings
-        run: ruff format --check
+  mypy:
+    needs: changes
+    runs-on: ubuntu-latest
+    if: github.event.pull_request.draft == false && needs.changes.outputs.mypy
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: '3.13.1'
+      - run: pip install tox==4.26.0
+      - name: Check Python Package structure
+        run: |
+          tox -re mypy
+
+  package:
+    needs: changes
+    runs-on: ubuntu-latest
+    if: github.event.pull_request.draft == false && needs.changes.outputs.lint
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: '3.13.1'
+      - run: pip install tox
+      - name: Check Python Package structure
+        run: tox -e pkg_meta
