+++ .github/workflows/release.yml
@@ -46,7 +46,6 @@ jobs:
         uses: pypa/gh-action-pypi-publish@v1.12.4
         with:
           attestations: true
-
   dockerize:
     runs-on: ubuntu-latest
     permissions:
