+++ docker/Dockerfile
@@ -1,6 +1,7 @@
 FROM python:3.12-slim-bookworm AS base_os
 ARG VERSION
-ARG DISTDIR
+ARG GIT_SHA
+ARG BUILD_DATE
 
 RUN set -x \
   && runtimeDeps=" \
@@ -17,10 +18,10 @@ wget \
   && apt-get purge -y --auto-remove $buildDeps
 
 RUN groupadd --gid 1024 os4d \
-    && adduser --disabled-login --disabled-password --no-create-home --ingroup os4d -q brd
+    && adduser --disabled-login --disabled-password --no-create-home --ingroup os4d -q birder
 
-# ------- builder -------
-FROM base_os AS builder
+# ------- deps -------
+FROM base_os AS deps
 RUN set -x \
     && buildDeps="build-essential \
 cmake \
@@ -44,33 +45,44 @@ zlib1g-dev  \
     && apt-get clean \
     && rm -rf /var/lib/apt/lists/*
 
-RUN pip install uwsgi
+RUN pip install uwsgi uv
+
+# ------- builder -------
+FROM deps AS builder
+ENV VIRTUAL_ENV=/venv \
+    UV_PROJECT_ENVIRONMENT=/venv
+
+RUN <<EOF cat> /RELEASE
+{"version": "$VERSION",
+  "commit": "$GIT_SHA"
+  "date": "$BUILD_DATE"
+}
+EOF
+WORKDIR /code
+COPY . /code
+
+RUN ls -al /code/src/
+
+RUN uv venv $VIRTUAL_ENV \
+    && uv sync --no-dev --frozen --no-editable
 
 FROM base_os AS dist
-COPY --from=builder /usr/local/bin/uwsgi /usr/local/bin/
+ADD docker/conf /conf/
+ADD docker/bin/* /usr/local/bin/
 
-ENV PATH=/venv/bin:/usr/local/bin/:/usr/bin:/bin \
+COPY --from=builder /usr/local/bin/uwsgi /usr/local/bin/
+COPY --from=builder /venv /venv
+
+ENV PATH=/venv/bin:/usr/local/bin/:/usr/bin:/bin:/usr/sbin \
     VERSION=$VERSION \
     PYTHONUNBUFFERED=1 \
     PYTHONDONTWRITEBYTECODE=1 \
     DJANGO_SETTINGS_MODULE="birder.config.settings" \
     STATIC_URL="/static/" \
     PYTHONPATH=""\
+    GIT_SHA=$GIT_SHA \
     PGSSLCERT="/tmp/postgresql.crt" \
-    UV_PROJECT_ENVIRONMENT=/venv \
-    VIRTUAL_ENV=/venv \
     UWSGI_PROCESSES=4
-
-RUN <<EOF cat> /RELEASE
-{"version": "$VERSION",
-}
-EOF
-WORKDIR /app
-ADD docker/conf /conf/
-ADD docker/bin/* /usr/local/bin/
-
-RUN python3 -m venv /venv \
-    && /venv/bin/pip install --find-links=/conf/ os4d_birder==${VERSION}
 
 EXPOSE 8000
 ENTRYPOINT ["docker-entrypoint.sh"]
@@ -79,11 +91,13 @@ CMD ["run"]
 
 LABEL maintainer="brd@os4d.org"
 LABEL org.opencontainers.image.authors="brd@os4d.org"
-LABEL org.opencontainers.image.description="Birder"
+LABEL org.opencontainers.image.created="$BUILD_DATE"
+LABEL org.opencontainers.image.description="Birder: Open source service uptime monitor"
 LABEL org.opencontainers.image.documentation="https://github.com/os4d/birder/"
-LABEL org.opencontainers.image.title="brd Country Workspace"
-LABEL org.opencontainers.image.licenses="https://github.com/os4d/birder/blob/master/LICENSE"
-LABEL org.opencontainers.image.source="https://github.com/os4d/birder"
-LABEL org.opencontainers.image.url="https://hub.docker.com/repository/docker/os4d/bbirder/tags/"
-LABEL org.opencontainers.image.vendor="os4d"
+LABEL org.opencontainers.image.licenses="https://github.com/os4d/birder/blob/${GIT_SHA:-develop}/LICENSE.md"
+LABEL org.opencontainers.image.revision=$GIT_SHA
+LABEL org.opencontainers.image.source="https://github.com/os4d/birder/tree/${GIT_SHA:-develop}/"
+LABEL org.opencontainers.image.title="Birder"
 LABEL org.opencontainers.image.version="$VERSION"
+LABEL org.opencontainers.image.url="https://hub.docker.com/repository/docker/os4d/bbirder/tags/"
+LABEL org.opencontainers.image.vendor="OS4D"
